⚡️ :: NotificationDTO 분할
