+++ Service/Sources/Data/RemoteDataSource/RestAPI/DataMapping/NotificationListDTO.swift
@@ -1,5 +1,3 @@
-// swiftlint:disable nesting
-
 import Foundation
 
 // MARK: - Data Transfer Object
@@ -7,24 +5,5 @@ struct NotificationListDTO: Decodable {
     private enum CodingKeys: String, CodingKey {
         case notificationList = "notification_list"
     }
-    let notificationList: [NotificationInformation]
-}
-
-extension NotificationListDTO {
-    struct NotificationInformation: Decodable {
-        private enum CodingKeys: String, CodingKey {
-            case id
-            case title
-            case content
-            case type
-            case value
-            case read = "is_read"
-        }
-        let id: Int
-        let title: String
-        let content: String
-        let type: String
-        let value: Int
-        let read: Bool
-    }
+    let notificationList: [NotificationDTO]
 }
